July 15th:
	- Converted Simulation Time from Ms to us.
	- Converted Time type from Double to Float
	- Converted Bandwidth type from Integer to Float

+++ Simulator/src/entities/controllers/routers/Dijkstra.java
@@ -12,13 +12,13 @@ public class Dijkstra extends Router {
 	private HashMap<Integer, SDNSwitch> nodes;
 	private HashMap<Integer, Integer> unvisited; // <SwitchID, sameSwitchID>
 	private HashMap<Integer, Integer> previous; // <SwitchID, SwitchID>
-	private HashMap<Integer, Double> distance; // <SwitchID, PropagationDelay>
+	private HashMap<Integer, Float> distance; // <SwitchID, PropagationDelay>
 	private HashMap<Integer, Link> result; // <SwitchID, Link>
 
 	/* Constructor */
 	public Dijkstra(HashMap<Integer, SDNSwitch> nodes) {
 		this.nodes = nodes;
-		distance = new HashMap<Integer, Double>();
+		distance = new HashMap<Integer, Float>();
 		previous = new HashMap<Integer, Integer>();
 		unvisited = new HashMap<Integer, Integer>();
 		result = new HashMap<Integer, Link>();
@@ -27,23 +27,23 @@ public class Dijkstra extends Router {
 
 	/* Objective::Finding the optimal paths for each Node */
 	public HashMap<Integer, Link> run(int srcID, int targetID) {
-		distance = new HashMap<Integer, Double>();
+		distance = new HashMap<Integer, Float>();
 		previous = new HashMap<Integer, Integer>();
 		unvisited = new HashMap<Integer, Integer>();
 
 		/* Initialization */
 		for (SDNSwitch currentNode : nodes.values()) {
-			distance.put(currentNode.getID(), Double.MAX_VALUE); // Unknown distance from source to v
+			distance.put(currentNode.getID(), Float.MAX_VALUE); // Unknown distance from source to v
 			previous.put(currentNode.getID(), srcID); // Previous node in optimal path from source
 			unvisited.put(currentNode.getID(), currentNode.getID()); // All nodes initially in Q (unvisited nodes)
 		}
 
-		distance.put(srcID, 0.0); // Distance from source to source
+		distance.put(srcID, (float) 0); // Distance from source to source
 
 		while (!unvisited.isEmpty()) {
 			int minNodeID = -1;
 
-			Double minValue = Double.MAX_VALUE;
+			float minValue = Float.MAX_VALUE;
 			for (int nodeID : unvisited.keySet()) { // Node with the least distance
 				if (distance.get(nodeID) < minValue) { // will be selected first
 					minValue = distance.get(nodeID);
@@ -53,7 +53,7 @@ public class Dijkstra extends Router {
 			unvisited.remove(minNodeID);
 
 			for (int n : nodes.get(minNodeID).networkLinks.keySet()) { // For each neighbor n of minNode
-				Double alt = distance.get(minNodeID) + nodes.get(minNodeID).networkLinks.get(n).getPropagationDelay();
+				float alt = distance.get(minNodeID) + nodes.get(minNodeID).networkLinks.get(n).getPropagationDelay();
 				if (alt < distance.get(n)) {
 					distance.put(n, alt);
 					previous.put(n, minNodeID);
